chatbot/requirements.txt
langchain_core
langchain_openai
streamlit
openai
python-dotenv
langchain_community
sse_servellete